+++ backend/kafka-queue/types.go
@@ -67,8 +67,9 @@ type AddSessionPropertiesArgs struct {
 	PropertiesObject interface{}
 }
 type PushBackendPayloadArgs struct {
-	SessionSecureID string
-	Errors          []*customModels.BackendErrorObjectInput
+	ProjectVerboseID *string
+	SessionSecureID  *string
+	Errors           []*customModels.BackendErrorObjectInput
 }
 
 type PushMetricsArgs struct {
@@ -79,8 +80,9 @@ type PushMetricsArgs struct {
 }
 
 type MarkBackendSetupArgs struct {
-	SecureID  string
-	ProjectID int
+	SecureID         *string
+	ProjectID        int
+	ProjectVerboseID *string
 }
 
 type AddSessionFeedbackArgs struct {