+++ sdk/highlight-node/src/client.ts
@@ -44,6 +44,7 @@ export class Highlight {
 	metrics: Array<InputMaybe<MetricInput>> = []
 	lastBackendSetupEvent: number = 0
 	_errorContext: ErrorContext | undefined
+	_projectID: string
 
 	constructor(options: NodeOptions) {
 		this._backendUrl = options.backendUrl || 'https://pub.highlight.run'
@@ -60,6 +61,7 @@ export class Highlight {
 				sourceContextCacheSizeMB: options.errorSourceContextCacheSizeMB,
 			})
 		}
+		this._projectID = options.projectID
 	}
 
 	recordMetric(
@@ -82,8 +84,8 @@ export class Highlight {
 
 	consumeCustomError(
 		error: Error,
-		secureSessionId: string,
-		requestId: string,
+		secureSessionId: string | undefined,
+		requestId: string | undefined,
 	) {
 		let res: StackFrameWithSource[] = []
 		try {
@@ -123,12 +125,13 @@ export class Highlight {
 		})
 	}
 
-	consumeCustomEvent(secureSessionId: string) {
+	consumeCustomEvent(secureSessionId?: string) {
 		const sendBackendSetup =
 			Date.now() - this.lastBackendSetupEvent > this.BACKEND_SETUP_TIMEOUT
 		if (sendBackendSetup) {
 			this._graphqlSdk
 				.MarkBackendSetup({
+					project_id: this._projectID,
 					session_secure_id: secureSessionId,
 				})
 				.then(() => {
@@ -145,6 +148,7 @@ export class Highlight {
 			return
 		}
 		const variables: PushBackendPayloadMutationVariables = {
+			project_id: this._projectID,
 			errors: this.errors,
 		}
 		this.errors = []